+++ CHANGELOG.md
@@ -8,11 +8,11 @@
 
 ## v0.2.0 (2022-1-15)
 
-- REnum.Enumerable.ActiveSupport
+- REnum.ActiveSupport
 
 ## v0.1.0 (2022-1-14)
 
 - Initial release.
 - Support all native enumerable functions.
-- REnum.Enumerable.Ruby
-- REnum.Enumerable.Support
+- REnum.Ruby
+- REnum.Support